scripts/requirements.txt
# Dependencias para scripts de sincronización DELIA
# Instalar con: pip install -r scripts/requirements.txt

# Cliente de Supabase
supabase>=2.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0

# Cliente de Google Sheets API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.0.0

# Validación y transformación de datos
jsonschema>=4.20.0

# Utilidades adicionales
requests>=2.31.0

